requirements.txt
# ── Data ──────────────────────────────────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=14.0.0          # Parquet read/write
pandas-ta>=0.3.14b       # 130+ technical indicators, pure Python (no C deps)
requests>=2.31.0
python-dotenv>=1.0.0

# ── ML — Tree Models ──────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.1.0
optuna>=3.5.0            # hyperparameter tuning
hmmlearn>=0.3.0          # Hidden Markov Model for regime detection

# ── ML — Deep Learning ────────────────────────────────────────────────────────
# Install PyTorch with CUDA 11.8 separately (see SETUP.md)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torch>=2.1.0
pytorch-forecasting>=1.0.0   # Temporal Fusion Transformer
pytorch-lightning>=2.1.0
transformers>=4.36.0         # FinBERT for news sentiment

# ── Explainability ────────────────────────────────────────────────────────────
shap>=0.44.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.1

# ── Model Storage ─────────────────────────────────────────────────────────────
huggingface-hub>=0.20.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.31.0
plotly>=5.18.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
joblib>=1.3.0
scipy>=1.12.0
statsmodels>=0.14.0      # statistical tests

# ── News / NLP ────────────────────────────────────────────────────────────────
newsapi-python>=0.2.7
